+++ src/sql/magic_plot.py
@@ -4,7 +4,7 @@ from IPython.core.magic import (
     magics_class,
 )
 from IPython.core.magic_arguments import argument, magic_arguments
-
+from ploomber_core.exceptions import modify_exceptions
 
 try:
     from traitlets.config.configurable import Configurable
@@ -49,6 +49,7 @@ class SqlPlotMagic(Magics, Configurable):
         action="append",
         dest="with_",
     )
+    @modify_exceptions
     def execute(self, line="", cell="", local_ns=None):
         """
         Plot magic