+++ src/sql/store.py
@@ -1,5 +1,6 @@
 from typing import Iterator, Iterable
 from collections.abc import MutableMapping
+from ploomber_core.exceptions import modify_exceptions
 
 from jinja2 import Template
 
@@ -54,6 +55,7 @@ class SQLStore(MutableMapping):
         # TODO: if with is false, WITH should not appear
         return SQLQuery(self, query, with_)
 
+    @modify_exceptions
     def store(self, key, query, with_=None):
         if with_ and key in with_:
             raise ValueError(f"Script name ({key!r}) cannot appear in with_ argument")