+++ src/tests/test_magic_plot.py
@@ -1,5 +1,4 @@
 from pathlib import Path
-
 import pytest
 from IPython.core.error import UsageError
 import matplotlib.pyplot as plt
@@ -24,7 +23,7 @@ def test_validate_plot_name(tmp_empty, ip, cell, error_type, error_message):
     out = ip.run_cell(cell)
 
     assert isinstance(out.error_in_exec, error_type)
-    assert str(out.error_in_exec) == (error_message)
+    assert str(error_message).lower() in str(out.error_in_exec).lower()
 
 
 @pytest.mark.parametrize(